refactor(export): split merged-exporter's RelAggregates dedup into a sibling module
CI's module-size ratchet pins merged-exporter.ts at exactly 1667 lines
(main was already at the budget with zero headroom); the partial-redundancy
fix pushed it to 1709. Extract skipRedundantRelAggregates and its render-time
counterpart into merged-rel-aggregates.ts rather than raising the budget.
findEntitiesByType/extractStepAttribute stay MergedExporter's own methods
(the single source of truth) and are passed in as parameters, so no logic is
duplicated. merged-exporter.ts is now 1649 lines (18 lines of headroom).
